requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyPDF2
python-multipart
beautifulsoup4
google-generativeai